refactor: change the structure of router, components and pages to have scalable design
