+++ src/components/Movies/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./Movies";
