+++ src/layout/Layout.tsx
@@ -1,0 +1,15 @@
+import { Outlet } from "react-router-dom";
+import Header from "@/components/Header";
+
+function Layout() {
+  return (
+    <div className="App">
+      <Header />
+      <div className="container">
+        <Outlet />
+      </div>
+    </div>
+  );
+}
+
+export default Layout;
