+++ src/pages/Starred/Starred.tsx
@@ -1,0 +1,11 @@
+import StarredComponent from "@/components/Starred";
+
+function Starred() {
+  return (
+    <div>
+      <StarredComponent viewTrailer={{}} />
+    </div>
+  );
+}
+
+export default Starred;
