+++ src/pages/Watchlater/Watchlater.tsx
@@ -1,0 +1,11 @@
+import WatchLaterComponent from "@/components/Watchlater";
+
+function Watchlater() {
+  return (
+    <div>
+      <WatchLaterComponent viewTrailer={{}} />
+    </div>
+  );
+}
+
+export default Watchlater;
